(PC-5508) Check model is aligned with database in CI
