+++ .env.testauto
@@ -1,5 +1,6 @@
 ADMINISTRATION_EMAIL_ADDRESS=administration@example.com
 DEV_EMAIL_ADDRESS=dev@example.com
+DATABASE_URL=postgresql://pytest:pytest@localhost:5432/pass_culture
 SUPPORT_EMAIL_ADDRESS=support@example.com
 DEMARCHES_SIMPLIFIEES_ENROLLMENT_PROCEDURE_ID=2567158
 DEMARCHES_SIMPLIFIEES_ENROLLMENT_PROCEDURE_ID_v2=6712558
